UI - Rename 'supervision' plugin to 'monitoring'
