+++ setup.py
@@ -26,5 +26,5 @@ setup(
         'share/ssl/temboard_ca_certs_CHANGEME.pem',
         'share/temboard.conf.sample',
         'share/temboard.logrotate',
-        'temboardui/plugins/supervision/sql/supervision.sql'
+        'temboardui/plugins/monitoring/sql/monitoring.sql'
     ])])
