+++ share/sql/dev-fixture.sql
@@ -50,7 +50,7 @@ SELECT agent_address, agent_port, plug
         UNION ALL
         SELECT 'dashboard'
         UNION ALL
-        SELECT 'supervision'
+        SELECT 'monitoring'
         UNION ALL
         SELECT 'settings'
     ) s(plug);
