+++ temboard.dev.conf
@@ -13,9 +13,9 @@ ssl_ca_cert_file = share/ssl/temboard_ca_certs_CHANGEME.pem
 # Cookie secret key
 cookie_secret = SECRETKEYTOBECHANGED
 # Plugins
-plugins = ["dashboard", "settings", "activity", "supervision"]
+plugins = ["dashboard", "settings", "activity", "monitoring"]
 # Array of plugins requiring a PG connexion to the repository.
-plugins_orm_engine = ["supervision"]
+plugins_orm_engine = ["monitoring"]
 # Working dir
 home = .
