+++ temboardui/configuration.py
@@ -28,8 +28,8 @@ class Configuration(configparser.ConfigParser):
             'ssl_key_file': None,
             'ssl_ca_cert_file': None,
             'cookie_secret': None,
-            'plugins': ["dashboard", "settings", "activity", "supervision"],
-            'plugins_orm_engine': ["supervision"],
+            'plugins': ["dashboard", "settings", "activity", "monitoring"],
+            'plugins_orm_engine': ["monitoring"],
             'home': '/var/run/temboard'
         }
         self.logging = {
