+++ temboardui/plugins/monitoring/templates/menu.html
@@ -1,0 +1,14 @@
+<ul class="nav nav-tabs">
+  <li class="{{'active' if period == 'day' else ''}}">
+    <a href="/server/{{instance.agent_address}}/{{instance.agent_port}}/monitoring/day">Last day</a>
+  </li>
+  <li class="{{'active' if period == 'week' else ''}}">
+    <a href="/server/{{instance.agent_address}}/{{instance.agent_port}}/monitoring/week">Last week</a>
+  </li>
+  <li class="{{'active' if period == 'month' else ''}}">
+    <a href="/server/{{instance.agent_address}}/{{instance.agent_port}}/monitoring/month">Last month</a>
+  </li>
+  <li class="{{'active' if period == 'year' else ''}}">
+    <a href="/server/{{instance.agent_address}}/{{instance.agent_port}}/monitoring/year">Last year</a>
+  </li>
+</ul>
