--- temboardui/plugins/supervision/templates/menu.html
@@ -1,14 +1,0 @@
-<ul class="nav nav-tabs">
-  <li class="{{'active' if period == 'day' else ''}}">
-    <a href="/server/{{instance.agent_address}}/{{instance.agent_port}}/supervision/day">Last day</a>
-  </li>
-  <li class="{{'active' if period == 'week' else ''}}">
-    <a href="/server/{{instance.agent_address}}/{{instance.agent_port}}/supervision/week">Last week</a>
-  </li>
-  <li class="{{'active' if period == 'month' else ''}}">
-    <a href="/server/{{instance.agent_address}}/{{instance.agent_port}}/supervision/month">Last month</a>
-  </li>
-  <li class="{{'active' if period == 'year' else ''}}">
-    <a href="/server/{{instance.agent_address}}/{{instance.agent_port}}/supervision/year">Last year</a>
-  </li>
-</ul>
