+++ temboardui/static/css/temboard.css
@@ -282,7 +282,7 @@ td {
 	color: #555;
 }
 
-.supervision-chart {
+.monitoring-chart {
 	width: 100%;
 	height: 300px;
 }
